refactor: Rename Executor -> Rebalancer
