+++ typescript/cli/src/rebalancer/rebalancer/WithSemaphore.test.ts
@@ -3,21 +3,21 @@ import chaiAsPromised from 'chai-as-promised';
 import Sinon from 'sinon';
 
 import { Config } from '../config/Config.js';
-import { IExecutor } from '../interfaces/IExecutor.js';
+import { IRebalancer } from '../interfaces/IRebalancer.js';
 import { RebalancingRoute } from '../interfaces/IStrategy.js';
 
 import { WithSemaphore } from './WithSemaphore.js';
 
 chai.use(chaiAsPromised);
 
-class MockExecutor implements IExecutor {
+class MockRebalancer implements IRebalancer {
   rebalance(_routes: RebalancingRoute[]): Promise<void> {
     return Promise.resolve();
   }
 }
 
 describe('WithSemaphore', () => {
-  it('should call the underlying executor', async () => {
+  it('should call the underlying rebalancer', async () => {
     const config = {
       chains: {
         chain1: {
@@ -32,9 +32,9 @@ describe('WithSemaphore', () => {
       } as any as RebalancingRoute,
     ];
 
-    const executor = new MockExecutor();
-    const rebalanceSpy = Sinon.spy(executor, 'rebalance');
-    const withSemaphore = new WithSemaphore(config, executor);
+    const rebalancer = new MockRebalancer();
+    const rebalanceSpy = Sinon.spy(rebalancer, 'rebalance');
+    const withSemaphore = new WithSemaphore(config, rebalancer);
     await withSemaphore.rebalance(routes);
 
     expect(rebalanceSpy.calledOnce).to.be.true;
@@ -50,9 +50,9 @@ describe('WithSemaphore', () => {
       },
     } as any as Config;
 
-    const executor = new MockExecutor();
-    const rebalanceSpy = Sinon.spy(executor, 'rebalance');
-    const withSemaphore = new WithSemaphore(config, executor);
+    const rebalancer = new MockRebalancer();
+    const rebalanceSpy = Sinon.spy(rebalancer, 'rebalance');
+    const withSemaphore = new WithSemaphore(config, rebalancer);
     await withSemaphore.rebalance([]);
 
     expect(rebalanceSpy.calledOnce).to.be.false;
@@ -73,9 +73,9 @@ describe('WithSemaphore', () => {
       } as any as RebalancingRoute,
     ];
 
-    const executor = new MockExecutor();
-    const rebalanceSpy = Sinon.spy(executor, 'rebalance');
-    const withSemaphore = new WithSemaphore(config, executor);
+    const rebalancer = new MockRebalancer();
+    const rebalanceSpy = Sinon.spy(rebalancer, 'rebalance');
+    const withSemaphore = new WithSemaphore(config, rebalancer);
     await withSemaphore.rebalance(routes);
 
     expect(rebalanceSpy.calledOnce).to.be.true;
@@ -98,8 +98,8 @@ describe('WithSemaphore', () => {
       } as any as RebalancingRoute,
     ];
 
-    const executor = new MockExecutor();
-    const withSemaphore = new WithSemaphore(config, executor);
+    const rebalancer = new MockRebalancer();
+    const withSemaphore = new WithSemaphore(config, rebalancer);
 
     await expect(withSemaphore.rebalance(routes)).to.be.rejectedWith(
       "Cannot read properties of undefined (reading 'bridgeLockTime')",
@@ -121,9 +121,9 @@ describe('WithSemaphore', () => {
       } as any as RebalancingRoute,
     ];
 
-    const executor = new MockExecutor();
-    const rebalanceSpy = Sinon.spy(executor, 'rebalance');
-    const withSemaphore = new WithSemaphore(config, executor);
+    const rebalancer = new MockRebalancer();
+    const rebalanceSpy = Sinon.spy(rebalancer, 'rebalance');
+    const withSemaphore = new WithSemaphore(config, rebalancer);
 
     const rebalancePromise1 = withSemaphore.rebalance(routes);
     const rebalancePromise2 = withSemaphore.rebalance(routes);
